Enable pipefail
Slither is now invoked as part of a series of piped commands. However
the default shell behavior is to mask any error reported if the last
command succeeds. This causes failures in slither to be masked
unintentionally. Enable pipefail to make the pipeline of commands
fail instead, restoring the previous behavior.